refactoring code for parser generator
