+++ apps/wit/wit.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+int main(int argc, char *argv[]) {
+    // Your code here
+    
+    return 0;
+}
