--- include/parser/amodule.h
@@ -1,27 +1,0 @@
-/*
- * amodule.h
- *
- * Copyright (C) 2021 Ligang Wang <ligangwangs@gmail.com>
- *
- * one module is one translation unit
- */
-#ifndef __MLANG_AMODULE_H__
-#define __MLANG_AMODULE_H__
-
-#include "parser/ast.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-struct amodule {
-    struct ast_node *root_ast;
-    const char *code;
-};
-
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
