+++ include/parser/m_parsing_table.h
@@ -8,50 +8,15 @@
 #define __MLANG_M_PARSING_TABLE_H__
 
 #include "clib/typedef.h"
-#include "parser/grammar.h"
 #include "parser/node_type.h"
+#include "pgen/lalr_parser_generator.h"
+#include "parser/m_parsing_table_def.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#include "parser/m_parsing_table_def.h"
 
-#define MAX_SYMBOLS_RULE 16
-
-struct rule_action {
-    enum node_type node_type;
-    u8 item_index[MAX_SYMBOLS_RULE]; // 0: is the first exp item value parsed at right side of grammar rule
-    u8 item_index_count;
-};
-
-enum action_code {
-    E, // indicator to do error recoverage
-    S, // shift the state
-    R, // reduce with production rule
-    A, // complete the parsing successfully
-    G, // goto for state i, and X nonterm
-};
-
-struct parser_action {
-    enum action_code code;
-    union {
-        u16 state_index; // next state index if action is shift
-        u16 rule_index; // index of production rule if action is reduce
-    };
-};
-
-
-// converted grammer rule with integer
-struct parse_rule {
-    const char *rule_string;
-    u16 lhs; // non terminal symbol index
-    u16 rhs[MAX_SYMBOLS_RULE]; // right hand side of production rule
-    u8 symbol_count; // right side of
-    struct rule_action action;
-};
-
-#define MAX_KERNEL_ITEMS   11
 struct parse_rule_state{
     u8 rule;    //rule index
     u8 dot;     //dot position
