--- include/parser/parser_def.h
@@ -1,24 +1,0 @@
-/*
- * parser_def.h
- * 
- * Copyright (C) 2022 Ligang Wang <ligangwangs@gmail.com>
- *
- * header file for some macro definitions
- */
-#ifndef __MLANG_PARSER_DEF_H__
-#define __MLANG_PARSER_DEF_H__
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#define MAX_STATES 512
-#define MAX_RULES 256
-#define MAX_AUGMENTED_RULES 8192
-
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
